Refactor task orchestration into strategy and scheduler
