+++ test/multi-cgra/taskflow/resnet/simple_resnet_tosa.mlir
@@ -18,7 +18,7 @@
 // RUN: mlir-neura-opt %t.affine.mlir --affine-loop-tree-serialization \
 // RUN: --convert-affine-to-taskflow \
 // RUN: --construct-hyperblock-from-task \
-// RUN: '--orchestrate-task-on-cgra=orchestration-mode=spatial-temporal' \
+// RUN: '--orchestrate-task-on-cgra=scheduling-mode=spatial-temporal' \
 // RUN: --architecture-spec=%S/../../../arch_spec/architecture_4x4.yaml \
 // RUN: -o %t.map_4x4_spatial_temporal.mlir
 // RUN: FileCheck %s --input-file=%t.map_4x4_spatial_temporal.mlir --check-prefixes=MAP-SPATIAL-TEMPORAL-4x4
@@ -26,7 +26,7 @@
 // RUN: mlir-neura-opt %t.affine.mlir --affine-loop-tree-serialization \
 // RUN: --convert-affine-to-taskflow \
 // RUN: --construct-hyperblock-from-task \
-// RUN: '--orchestrate-task-on-cgra=orchestration-mode=spatial-temporal' \
+// RUN: '--orchestrate-task-on-cgra=scheduling-mode=spatial-temporal' \
 // RUN: --architecture-spec=%S/../../../arch_spec/architecture_with_counter.yaml \
 // RUN: -o %t.map_1x1_spatial_temporal.mlir
 // RUN: FileCheck %s --input-file=%t.map_1x1_spatial_temporal.mlir --check-prefixes=MAP-SPATIAL-TEMPORAL-1x1
@@ -34,7 +34,7 @@
 // RUN: mlir-neura-opt %t.affine.mlir --affine-loop-tree-serialization \
 // RUN: --convert-affine-to-taskflow \
 // RUN: --construct-hyperblock-from-task \
-// RUN: '--orchestrate-task-on-cgra=orchestration-mode=spatial-temporal' \
+// RUN: '--orchestrate-task-on-cgra=scheduling-mode=spatial-temporal' \
 // RUN: --architecture-spec=%S/../../../arch_spec/architecture_1x2.yaml \
 // RUN: -o %t.map_1x2_spatial_temporal.mlir
 // RUN: FileCheck %s --input-file=%t.map_1x2_spatial_temporal.mlir --check-prefixes=MAP-SPATIAL-TEMPORAL-1x2
